vite-scripts: Change injection point of entry path in HTML
The default setting, depending on the order of how plugins are
configured, would result in the HMR plugin code being injected AFTER the
entry script. This would mean that component code would execute before
HMR code was set up, and would fail.

Potentially there are 3 options to fix this;
1) change `enforce: 'pre'` to `post` to allow executing after other
   plugins
2) change the order of plugins in `scripts/react-vite/config` to append
   the `react-vite` plugin before the super plugins, but that could have
   other side-effects in the future.
3) change the `injectTo` from `prepend-head` to `head` to make sure it's
   always injected into other

Here the 3rd option is chosen.