muban: remove additionalData from scss-loader
When the @use directive is used in scss files, it must come before any
other rules, which is not possible when these imports exist as a
"prepend" to all files.

Instead, we now explicitly have to import these utilities in our scss
files, which is mostly handled by adding them to our plop template.