twig-server: Change FileSystem loader to use absolute paths
Relative paths are not officially supported by Twig running in PHP, so
we had to replace it with an absolute FileSystem loader.

With this change, support for namespaces was also added, and with this
new loader, we can automatically support multiple template dirs, which
makes that implementation easier.

Also created some utilities to parse and normalize the different ways
you can pass template paths.

And with the above, added Jest integration to test those utilities.

Template files also had to be renamed to .html.twig.