muban: move copy of twig files and server from webpack to command
The twig files and twig server were copied/generated in webpack, which
made it harder to execute them on other type of build actions.

This changes moves them to their own command, which can be executed as
part of any other build script.

It also means that storybook builds don't have to compile all pages to
get access to the twig files in the dist folder.